+++ engine/gameobject/src/gameobject/test/delete/test_gameobject_delete.cpp
@@ -151,7 +151,7 @@ TEST_F(DeleteTest, DeleteSelf)
         for (int i = 0; i < 512; ++i)
         {
             dmGameObject::HInstance go = dmGameObject::New(m_Collection, "/go.goc");
-            dmGameObject::SetPosition(go, Vectormath::Aos::Point3(i,i,i));
+            dmGameObject::SetPosition(go, Vectormath::Aos::Point3((float) i,(float) i, (float) i));
             ASSERT_NE((void*) 0, (void*) go);
             m_DeleteSelfInstances.push_back(go);
             m_DeleteSelfIndexToInstance[i] = go;
